+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Support named item lists ([#1665](../../pull/1665))
 - Add options to group within item lists ([#1666](../../pull/1666))
 - Make the filter field in item lists wider ([#1669](../../pull/1669))
+- Add a navigate option to item lists ([#1659](../../pull/1659))
 
 ### Changes